services/analytics-service/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
httpx
python-dotenv
numpy
scikit-learn
pika
